+++ src/interfaces/src/modules/interfaces.hpp
@@ -4,24 +4,12 @@
 
 #include "core/context.hpp"
 
+// module context
+#include "interfaces/context.hpp"
+
 // subs API
 #include "interfaces/oper.hpp"
 #include "interfaces/change.hpp"
-
-/**
- * Operational context for the interfaces module.
- */
-class InterfacesOperationalContext : public srpc::IModuleContext { };
-
-/**
- * Module changes context for the interfaces module.
- */
-class InterfacesModuleChangesContext : public srpc::IModuleContext { };
-
-/**
- * RPC context for the interfaces module.
- */
-class InterfacesRpcContext : public srpc::IModuleContext { };
 
 /**
  * @brief Interfaces module.
